strategy_flipster_python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "strategy-flipster"
version = "0.1.0"
description = "Flipster 선물 HFT 전략 — 멀티 거래소 시그널 기반"
requires-python = ">=3.10"
dependencies = [
    "pyzmq>=26.0",
    "websockets>=13.0",
    "httpx>=0.27",
    "uvloop>=0.21; sys_platform != 'win32'",
    "structlog>=24.0",
    "tomli>=2.0; python_version < '3.11'",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "mypy>=1.10",
]

[project.scripts]
strategy-flipster = "strategy_flipster.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/strategy_flipster"]

[tool.mypy]
strict = true
python_version = "3.12"

[tool.pytest.ini_options]
asyncio_mode = "auto"
